rename file, change ReactDOM on React, get data from backend
