+++ react-app/src/List.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Item from "./item";
+import Item from "./item.js";
 
 export default function List({ data }) {
   return (
