--- react-app/src/data/data.json
@@ -1,69 +1,0 @@
-[
-    {
-        "id": 5,
-        "nickname": "Player4",
-        "score": 4123,
-        "time": "00:04:12"
-      },
-      {
-        "id": 3,
-        "nickname": "Player3",
-        "score": 3412,
-        "time": "00:03:41"
-      },
-      {
-        "id": 2,
-        "nickname": "Player2",
-        "score": 2314,
-        "time": "00:02:34"
-      },
-      {
-        "id": 6,    
-        "nickname": "Player6",
-        "score": 2134,
-        "time": "00:01:43"
-      },
-      {
-        "id": 1,
-        "nickname": "Player1",
-        "score": 1234,
-        "time": "00:01:23"
-      },
-      {
-        "id": 5,
-        "nickname": "Player4",
-        "score": 4123,
-        "time": "00:04:12"
-      },
-      {
-        "id": 3,
-        "nickname": "Player3",
-        "score": 3412,
-        "time": "00:03:41"
-      },
-      {
-        "id": 2,
-        "nickname": "Player2",
-        "score": 2314,
-        "time": "00:02:34"
-      },
-      {
-        "id": 6,    
-        "nickname": "Player6",
-        "score": 2134,
-        "time": "00:01:43"
-      },
-      {
-        "id": 1,
-        "nickname": "Player1",
-        "score": 1234,
-        "time": "00:01:23"
-      },
-      {
-        "id": 2,
-        "nickname": "Maks",
-        "score": 12344,
-        "time": "00:01:23"
-      }
-  ]
-
