--- react-app/src/data/index.js
@@ -1,14 +1,0 @@
-import data from "./data.json";
-
-let currentData = [];
-
-function mapWithRank(row, i) {
-  return { ...row, ranking: i + 1 };
-}
-
-export function getInitialData() {
-  currentData = data.map(mapWithRank);
-  return currentData;
-}
-
-
