+++ react-app/src/users_item.js
@@ -1,0 +1,16 @@
+const actualData = await fetch(
+    `https://jsonplaceholder.typicode.com/users`  //paste URL from backend
+    ).then(response => response.json());
+  
+    
+  let currentData = [];
+  
+  function mapWithRank(row, i) {
+    return { ...row, ranking: i + 1 };
+  }
+  
+  export function getInitialData() {
+    currentData = actualData.map(mapWithRank);
+    return currentData;
+  }
+
